fixing db tests

+++ backend/app/routes/users.py
@@ -1,3 +1,4 @@
+from os import abort
 from flask import Blueprint, request, jsonify
 from ..models import db, User
 
@@ -18,12 +19,16 @@ def get_users():
 
 @bp.route('/<email>', methods=['GET'])
 def get_user(email):
-    user = User.query.get_or_404(email)
+    user = db.session.get(User, email)
+    if user is None:
+        abort(404)
     return jsonify({"email": user.email})
 
 @bp.route('/<email>', methods=['PUT'])
 def update_user(email):
-    user = User.query.get_or_404(email)
+    user = db.session.get(User, email)
+    if user is None:
+        abort(404)
     data = request.json
     user.email = data['email']
     db.session.commit()
@@ -31,7 +36,9 @@ def update_user(email):
 
 @bp.route('/<email>', methods=['DELETE'])
 def delete_user(email):
-    user = User.query.get_or_404(email)
+    user = db.session.get(User, email)
+    if user is None:
+        abort(404)
     db.session.delete(user)
     db.session.commit()
     return jsonify({"message": "User deleted successfully"})
